Part 1 of production
More complex than anticipated:
New class is needed to represent minion archetypes.
Masterboard needs an array of  these archetypes to compare against for purchases. Later, building a new minion will take this into account as well.
Then the propertyMenu can search the list for the requested purchase type and see if it exists and it can afford it.