Compie Production 2.0
Compie analyzes the map and bases production for each land mass on the total units there, and seeks to produce a proportionate number of inf. and tanks. It will only limit the number of inf. it builds, all other units it will build if it has money.
Changed speed factor to multiplier since it's an int.